+++ google/cloud/forseti/common/util/email/base_email_connector.py
@@ -69,71 +69,5 @@ class BaseEmailConnector(object):
         """
         pass
 
-    @staticmethod
-    def _add_recipients(email, email_recipients):
-        """Add multiple recipients to the sendgrid email object.
 
-        Args:
-            email (mail): Connector mail object
-            email_recipients (Str): comma-separated text of the email recipients
 
-        Returns:
-            email: Connector mail object with multiple recipients.
-        """
-        personalization = mail.Personalization()
-        recipients = email_recipients.split(',')
-        for recipient in recipients:
-            personalization.add_to(mail.Email(recipient))
-        email.add_personalization(personalization)
-        return email
-
-    @classmethod
-    def render_from_template(cls, template_file, template_vars):
-        """Fill out an email template with template variables.
-
-        Args:
-            template_file (str): The location of email template in filesystem.
-            template_vars (dict): The template variables to fill into the
-                template.
-
-        Returns:
-            str: The template content, rendered with the provided variables.
-        """
-        template_searchpath = os.path.abspath(
-            os.path.join(os.path.dirname(__file__), '../email_templates'))
-        template_loader = jinja2.FileSystemLoader(
-            searchpath=template_searchpath)
-        template_env = jinja2.Environment(loader=template_loader)
-        template = template_env.get_template(template_file)
-        return template.render(template_vars)
-
-    @classmethod
-    def create_attachment(cls, file_location, content_type, filename,
-                          disposition='attachment', content_id=None):
-        """Create a SendGrid attachment.
-
-        Email connector attachments file content must be base64 encoded.
-
-        Args:
-            file_location (str): The path of the file.
-            content_type (str): The content type of the attachment.
-            filename (str): The filename of attachment.
-            disposition (str): Content disposition, defaults to "attachment".
-            content_id (str): The content id.
-
-        Returns:
-            Attachment: A Connector Attachment.
-        """
-        file_content = ''
-        with open(file_location, 'rb') as f:
-            file_content = f.read()
-        content = base64.b64encode(file_content)
-
-        attachment = mail.Attachment()
-        attachment.content = content
-        attachment.type = content_type
-        attachment.filename = filename
-        attachment.disposition = disposition
-        attachment.content_id = content_id
-
-        return attachment
