+++ google/cloud/forseti/notifier/notifiers/inventory_summary.py
@@ -17,14 +17,15 @@ from googleapiclient.errors import HttpError
 
 # pylint: disable=line-too-long
 from google.cloud.forseti.common.util import date_time
-from google.cloud.forseti.common.util.email import sendgrid_connector
+from google.cloud.forseti.common.util.email.sendgrid_connector import EmailUtil
 from google.cloud.forseti.common.util import errors as util_errors
 from google.cloud.forseti.common.util import file_uploader
 from google.cloud.forseti.common.util import logger
 from google.cloud.forseti.common.util import string_formats
-from google.cloud.forseti.notifier.notifiers.base_notification import BaseNotification
+from google.cloud.forseti.notifier.notifiers.base_notification import \
+    BaseNotification
 from google.cloud.forseti.services.inventory.storage import InventoryIndex
-from google.cloud.forseti.common.util.email.base_email_connector import BaseEmailConnector
+from google.cloud.forseti.common.util.email.sendgrid_connector import EmailUtil
 
 # pylint: enable=line-too-long
 
@@ -117,8 +118,13 @@ class InventorySummary(object):
         email_summary_config = (
             self.notifier_config.get('inventory').get('email_summary'))
 
-        email_util = sendgrid_connector.EmailUtil(
-            email_summary_config.get('sendgrid_api_key'))
+        email_connector_config_auth = self.notifier_config\
+            .get('email_connector_config').get('auth')
+
+        email_util = EmailUtil(
+            email_summary_config.get('sender'),
+            email_summary_config.get('recipient'),
+            email_connector_config_auth)
 
         email_subject = 'Inventory Summary: {0}'.format(
             self.inventory_index_id)
@@ -131,7 +137,7 @@ class InventorySummary(object):
 
         gsuite_dwd_status = self._get_gsuite_dwd_status(summary_data)
 
-        email_content = BaseEmailConnector.render_from_template(
+        email_content = email_util.render_from_template(
             'inventory_summary.jinja',
             {'inventory_index_id': self.inventory_index_id,
              'timestamp': timestamp,
@@ -140,9 +146,11 @@ class InventorySummary(object):
              'details_data': details_data})
 
         try:
-            sendgrid_connector.send(
-                email_sender=email_summary_config.get('sender'),
-                email_recipient=email_summary_config.get('recipient'),
+            email_util.send(
+                email_sender=self.notifier_config.get('email_connector_config')
+                .get('sender'),
+                email_recipient=self.notifier_config.get('email_connector_config')
+                .get('recipient'),
                 email_subject=email_subject,
                 email_content=email_content,
                 content_type='text/html')
